requirements-dev.txt
-r requirements.txt

exchange-calendars
fsspec
black
mypy
pandas-stubs
pytest
pytest-cov
pytest-xdist
ruff
typer
